% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{mergeClassFiles}
\alias{mergeClassFiles}
\title{Merging class files that are located in the same folder}
\usage{
mergeClassFiles(inPath, filenames, outPath, outFilename, ...)
}
\arguments{
\item{inPath}{}

\item{filenames}{}

\item{outPath}{}

\item{outFilename}{}
}
\description{
Merging class files that are located in the same folder
}

