% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{splitClassFile}
\alias{splitClassFile}
\title{Splitting class files from one single R file}
\usage{
splitClassFile(inPath, filename, outPath, splitTag = "#--split ")
}
\arguments{
\item{inPath}{}

\item{filename}{}

\item{outPath}{}

\item{splitTag}{default '#--split '}
}
\description{
Splitting class files from one single R file
}

